requirements.txt
databricks-agents==0.15.0
databricks-langchain==0.1.1
databricks-sdk
databricks-vectorsearch>=0.44
docling>=2.15.0
gradio>=5.12.0
langchain-community==0.3.14
langchain-core==0.3.29
langgraph==0.2.62
mlflow==2.19.0
numpy>=1.26.0
openai>=1.35.0
pyyaml>=6.0.1
transformers>=4.42.0
